[Refactor:RainbowGrades] Student.cpp patch (#62)
This is student.cpp file patch by @KCony 
Mainly to address calculation of each normalized grade. 
" Grades should be compared by the normalized grade only, not normalized
grade multiplied by percentages.
Otherwise, a grade of 90 for the gradeable with percentage 0.1 will
always be considered lower than a grade of 60 for the gradeable with
percentage 0.2 ", which we don't want, if we were to drop lowest grade.
(gradeable)
Also additional logics implemented to handle extra credit accurately. 

Additionally when running patch to default sample course, it was failing
assert(sum_percentage <= 1.0)
mostly seemed to be precision issue with floating-point arithmetic.
Handled it by padding tolerance of 0.00001

Overall changes has be reviewed and tested.
However, it would be best if Barb runs it with previous term data to
check the behavior, given that there is slight logic change.

---------

Co-authored-by: Konstantin Kuzmin <kkouzmin@mail.ru>