[RainbowGrades:Feature] add display options (#66)
Currently in web-based customization page, we have display option, which
is acts like a switch to enable and disable contents that needs to be
displayed to students. There are currently 7 options, which are
available, but we will expand these features to 14 options. It is a
valid checkbox on the Web GUI, but RainbowGrades cannot take the new 7
options, and will silently break if the new options are checked. Hence,
this PR will fix it by adding a placeholder to handle additional tokens
in display options.