[Feature:RainbowGrades] Display Extensions in RainbowGrades (#63)
Previously, no information on a student's late day exceptions were
provided in rainbow grades. Now, all excused extension information is
collected and stored in each student object. A row was added to the main
rainbow grades table to display a student's total number of extensions,
and a breakout table was added underneath the main table to display
information about each individual extension. Additonally, a blue outline
was added to table cell with excused extensions and text giving
extension information is listed on hover.

Picture of changes to tables:

![image](https://github.com/Submitty/RainbowGrades/assets/143554378/fa8ab970-20bb-4382-a8c0-a1baf664297c)
Hover Feature:

![image](https://github.com/Submitty/RainbowGrades/assets/143554378/cf243ae3-5f67-4f19-b379-77f791e44a51)

---------

Co-authored-by: Jaeseok Kang <ziesski@gmail.com>
Co-authored-by: Barb Cutler <Barb Cutler>